+++ test/promise-readable-stream1.js
@@ -9,8 +9,8 @@ chai.use(chaiAsPromised)
 chai.should()
 
 Feature('Test promise-readable module with stream1 API', () => {
-  const PromiseReadable = require('../lib/promise-readable')
-  const EventEmitter = require('events')
+  const { PromiseReadable } = require('../lib/promise-readable')
+  const { EventEmitter } = require('events')
 
   class MockStream extends EventEmitter {
     constructor () {