+++ examples/read-all.js
@@ -1,12 +1,14 @@
 'use strict'
 
 const { PromiseReadable } = require('../lib/promise-readable')
-const fs = require('fs')
+const { createReadStream } = require('fs')
 
 async function main () {
-  const rstream = new PromiseReadable(fs.createReadStream(process.argv[2] || '/etc/hosts'))
+  const rstream = new PromiseReadable(createReadStream(process.argv[2] || '/etc/hosts'))
   const data = await rstream.readAll()
-  console.log(`Read ${data.length} bytes in total`)
+  if (data != null) {
+    console.log(`Read ${data.length} bytes in total`)
+  }
 }
 
 main().catch(console.error)