+++ lib/promise-readable.d.ts
@@ -3,8 +3,8 @@ export declare class PromiseReadable {
 
   constructor (stream: NodeJS.ReadableStream)
 
-  read (size?: number): Promise<Buffer>
-  readAll (): Promise<Buffer>
+  read (size?: number): Promise<Buffer | undefined>
+  readAll (): Promise<Buffer | undefined>
 
   once (event: 'close'): Promise<void>
   once (event: 'end'): Promise<void>